+++ tests/plugins/clipboard/dropimage.js
@@ -150,6 +150,47 @@
 			} );
 		},
 
+		// (#5431)
+		'test dropping image when the clipboard_handleImages configuration option is OFF displays notification about unsupported image type':
+			function() {
+				var originalClipboard_handleImages = CKEDITOR.config.clipboard_handleImages;
+				CKEDITOR.config.clipboard_handleImages = false;
+
+				var dropEvt = bender.tools.mockDropEvent(),
+					imageType = 'image/png',
+					expectedData = '<p class="p">Paste image here:</p>',
+					expectedMsgRegex  = prepareNotificationRegex( this.editor.lang.clipboard.fileFormatNotSupportedNotification ),
+					expectedDuration = this.editor.config.clipboard_notificationDuration,
+					notificationSpy = sinon.spy( this.editor, 'showNotification' );
+
+				FileReader.setFileMockType( imageType );
+				FileReader.setReadResult( 'load' );
+
+				setHtmlWithSelection( this.editor, '<p class="p">Paste image here:^</p>' );
+				assertImageDrop(  {
+					editor: this.editor,
+					event: dropEvt,
+					type: imageType,
+					expectedData: expectedData,
+					dropRange: {
+						dropContainer: this.editor.editable().findOne( '.p' ).getChild( 0 ),
+						dropOffset: 17
+					},
+					callback: function() {
+						notificationSpy.restore();
+
+						assert.areSame( 1, notificationSpy.callCount, 'There was only one notification' );
+						assert.isMatching( expectedMsgRegex, notificationSpy.getCall( 0 ).args[ 0 ],
+							'The notification had correct message' );
+						assert.areSame( 'info', notificationSpy.getCall( 0 ).args[ 1 ],
+						'The notification had correct type' );
+						assert.areSame( expectedDuration, notificationSpy.getCall( 0 ).args[ 2 ],
+							'The notification had correct duration' );
+					}
+				} );
+				CKEDITOR.config.clipboard_handleImages = originalClipboard_handleImages;
+			},
+
 		'test abort drop': function() {
 			var dropEvt = bender.tools.mockDropEvent(),
 				imageType = 'image/png',